Composite adjuster
